requirements.txt

psycopg2-binary==2.9.6
Faker==18.11.2
click==8.1.3
tabulate==0.9.0